% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_report}
\alias{clean_report}
\title{Deja solo las variables relevantes en la validación}
\usage{
clean_report(report, label, id)
}
\arguments{
\item{report}{Data frame}

\item{label}{Nombre de la columna que que se quiere limpiar}

\item{id}{Identificador de cada registro. Es un string}
}
\description{
Deja solo las variables relevantes en la validación
}
