% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_results}
\alias{plot_results}
\title{Valida la columna de un data frame, en base a reglas establecidas}
\usage{
plot_results(data, var)
}
\arguments{
\item{data}{Data frame}

\item{var}{String. Variable del data frame a ser graficada}
}
\description{
Valida una columna de un data frame, siguiendo las reglas establecidas con anterioridad.
}
